# HAMTv4 using generics instead of cbg.Deferred for perf gains