--- a/builtin/actor_tree.go
+++ b/builtin/actor_tree.go
@@ -9,6 +9,21 @@ import (
 	"golang.org/x/xerrors"
 )
 
+type ActorTree interface {
+	GetStore() adt.Store
+
+	Flush() (cid.Cid, error)
+	GetActorV4(addr address.Address) (*ActorV4, bool, error)
+	GetActorV5(addr address.Address) (*ActorV5, bool, error)
+	SetActorV4(addr address.Address, actor *ActorV4) error
+	SetActorV5(addr address.Address, actor *ActorV5) error
+	ForEachV4(fn func(addr address.Address, actor *ActorV4) error) error
+	ForEachV5(fn func(addr address.Address, actor *ActorV5) error) error
+	ForEachKey(fn func(addr address.Address) error) error
+}
+
+var _ ActorTree = (*actorTree)(nil)
+
 // Value type of the top level of the state tree.
 // Represents the on-chain state of a single actor.
 // This is the actor state for state tree version up to 4
@@ -29,42 +44,46 @@ type ActorV5 struct {
 }
 
 // A specialization of a map of ID-addresses to actor heads.
-type ActorTree struct {
+type actorTree struct {
 	Map   *adt.Map
 	Store adt.Store
 }
 
 // Initializes a new, empty state tree backed by a store.
-func NewTree(store adt.Store) (*ActorTree, error) {
+func NewTree(store adt.Store) (ActorTree, error) {
 	emptyMap, err := adt.MakeEmptyMap(store, DefaultHamtBitwidth)
 	if err != nil {
 		return nil, err
 	}
-	return &ActorTree{
+	return &actorTree{
 		Map:   emptyMap,
 		Store: store,
 	}, nil
 }
 
 // Loads a tree from a root CID and store.
-func LoadTree(s adt.Store, r cid.Cid) (*ActorTree, error) {
+func LoadTree(s adt.Store, r cid.Cid) (ActorTree, error) {
 	m, err := adt.AsMap(s, r, DefaultHamtBitwidth)
 	if err != nil {
 		return nil, err
 	}
-	return &ActorTree{
+	return &actorTree{
 		Map:   m,
 		Store: s,
 	}, nil
 }
 
+func (t *actorTree) GetStore() adt.Store {
+	return t.Store
+}
+
 // Writes the tree root node to the store, and returns its CID.
-func (t *ActorTree) Flush() (cid.Cid, error) {
+func (t *actorTree) Flush() (cid.Cid, error) {
 	return t.Map.Root()
 }
 
 // Loads the state associated with an address.
-func (t *ActorTree) GetActorV4(addr address.Address) (*ActorV4, bool, error) {
+func (t *actorTree) GetActorV4(addr address.Address) (*ActorV4, bool, error) {
 	if addr.Protocol() != address.ID {
 		return nil, false, xerrors.Errorf("non-ID address %v invalid as actor key", addr)
 	}
@@ -73,7 +92,7 @@ func (t *ActorTree) GetActorV4(addr address.Address) (*ActorV4, bool, error) {
 	return &actor, found, err
 }
 
-func (t *ActorTree) GetActorV5(addr address.Address) (*ActorV5, bool, error) {
+func (t *actorTree) GetActorV5(addr address.Address) (*ActorV5, bool, error) {
 	if addr.Protocol() != address.ID {
 		return nil, false, xerrors.Errorf("non-ID address %v invalid as actor key", addr)
 	}
@@ -83,14 +102,14 @@ func (t *ActorTree) GetActorV5(addr address.Address) (*ActorV5, bool, error) {
 }
 
 // Sets the state associated with an address, overwriting if it already present.
-func (t *ActorTree) SetActorV4(addr address.Address, actor *ActorV4) error {
+func (t *actorTree) SetActorV4(addr address.Address, actor *ActorV4) error {
 	if addr.Protocol() != address.ID {
 		return xerrors.Errorf("non-ID address %v invalid as actor key", addr)
 	}
 	return t.Map.Put(abi.AddrKey(addr), actor)
 }
 
-func (t *ActorTree) SetActorV5(addr address.Address, actor *ActorV5) error {
+func (t *actorTree) SetActorV5(addr address.Address, actor *ActorV5) error {
 	if addr.Protocol() != address.ID {
 		return xerrors.Errorf("non-ID address %v invalid as actor key", addr)
 	}
@@ -98,7 +117,7 @@ func (t *ActorTree) SetActorV5(addr address.Address, actor *ActorV5) error {
 }
 
 // Traverses all entries in the tree.
-func (t *ActorTree) ForEachV4(fn func(addr address.Address, actor *ActorV4) error) error {
+func (t *actorTree) ForEachV4(fn func(addr address.Address, actor *ActorV4) error) error {
 	var val ActorV4
 	return t.Map.ForEach(&val, func(key string) error {
 		addr, err := address.NewFromBytes([]byte(key))
@@ -109,7 +128,7 @@ func (t *ActorTree) ForEachV4(fn func(addr address.Address, actor *ActorV4) erro
 	})
 }
 
-func (t *ActorTree) ForEachV5(fn func(addr address.Address, actor *ActorV5) error) error {
+func (t *actorTree) ForEachV5(fn func(addr address.Address, actor *ActorV5) error) error {
 	var val ActorV5
 	return t.Map.ForEach(&val, func(key string) error {
 		addr, err := address.NewFromBytes([]byte(key))
@@ -121,7 +140,7 @@ func (t *ActorTree) ForEachV5(fn func(addr address.Address, actor *ActorV5) erro
 }
 
 // Traverses all keys in the tree, without decoding the actor states.
-func (t *ActorTree) ForEachKey(fn func(addr address.Address) error) error {
+func (t *actorTree) ForEachKey(fn func(addr address.Address) error) error {
 	return t.Map.ForEach(nil, func(key string) error {
 		addr, err := address.NewFromBytes([]byte(key))
 		if err != nil {