--- a/builtin/v10/migration/top.go
+++ b/builtin/v10/migration/top.go
@@ -33,11 +33,11 @@ func MigrateStateTree(ctx context.Context, store cbor.IpldStore, newManifestCID 
 	adtStore := adt9.WrapStore(ctx, store)
 
 	// Load input and output state trees
-	actorsIn, err := builtin.LoadTree(adtStore, actorsRootIn)
+	actorsIn, err := builtin.LoadLegacyTree(adtStore, actorsRootIn)
 	if err != nil {
 		return cid.Undef, xerrors.Errorf("loading state tree: %w", err)
 	}
-	actorsOut, err := builtin.NewTree(adtStore)
+	actorsOut, err := builtin.NewLegacyTree(adtStore)
 	if err != nil {
 		return cid.Undef, xerrors.Errorf("creating new state tree: %w", err)
 	}