--- a/builtin/v11/migration/top.go
+++ b/builtin/v11/migration/top.go
@@ -26,7 +26,7 @@ func MigrateStateTree(ctx context.Context, store cbor.IpldStore, newManifestCID 
 	adtStore := adt10.WrapStore(ctx, store)
 
 	// Load input and output state trees
-	actorsIn, err := builtin.LoadTree(adtStore, actorsRootIn)
+	actorsIn, err := builtin.LoadLegacyTree(adtStore, actorsRootIn)
 	if err != nil {
 		return cid.Undef, xerrors.Errorf("loading state tree: %w", err)
 	}