--- a/builtin/v12/migration/top.go
+++ b/builtin/v12/migration/top.go
@@ -25,7 +25,7 @@ func MigrateStateTree(ctx context.Context, store cbor.IpldStore, newManifestCID 
 	adtStore := adt11.WrapStore(ctx, store)
 
 	// Load input and output state trees
-	actorsIn, err := builtin.LoadTree(adtStore, actorsRootIn)
+	actorsIn, err := builtin.LoadLegacyTree(adtStore, actorsRootIn)
 	if err != nil {
 		return cid.Undef, xerrors.Errorf("loading state tree: %w", err)
 	}