--- a/builtin/v13/migration/top.go
+++ b/builtin/v13/migration/top.go
@@ -27,7 +27,7 @@ func MigrateStateTree(ctx context.Context, store cbor.IpldStore, newManifestCID 
 	adtStore := adt13.WrapStore(ctx, store)
 
 	// Load input and output state trees
-	actorsIn, err := builtin.LoadTree(adtStore, actorsRootIn)
+	actorsIn, err := builtin.LoadLegacyTree(adtStore, actorsRootIn)
 	if err != nil {
 		return cid.Undef, xerrors.Errorf("loading state tree: %w", err)
 	}