Ref: https://github.com/filecoin-project/go-state-types/issues/285

Continuing on the optimisation journey with the belief that allocations and GC are taking up a significant portion of the time, I  experimented with removing the `cbg.Deferred` use in go-hamt-ipld and using Go generics to replicate that functionality. When the HAMT doesn't know what value type it's got stored, it skips over the CBOR tokens until it finds a whole unit (single scalar token, whole nested object, whatever) and then slices the bytes for that value and passes it on in `cbg.Deferred` for the caller to decode themselves. This means double CBOR parsing and the allocation and GC of a byte slice that really shouldn't be there. So instead we can pass our serializable type down into the HAMT and it can just defer directly to it to decode its values. That work is in https://github.com/filecoin-project/go-hamt-ipld/pull/122 and this branch depends on that branch as `github.com/filecoin-project/go-hamt-ipld/v4`, keeping `/v3` intact.

Summary of changes so far:

* `builtin.ActorTree` is now an interface which only has the `ActorsV5` method variants and I've also made `builtin.LegacyActorTree` which has those plus the `ActorsV4` method variants.
* `builtin.NewLegacyTree` and `builtin.LoadLegacyTree` are the new names for the original form, these will give you what was always returned before.
* `GetStore()` on both interfaces, and `GetMap()` on `LegacyActorTree`—I need the latter for lotus, which only knows about `adt.Map` for now but with this change `adt.Map` can become a legacy concern and we can just use `hamt.Node[*ActorV5]` to get basically the same thing.
* `RunMigration` only takes an `ActorTree`, it doesn't need the v4 actors; it also uses `builtin.NewTree` to make the new v4 hamt with generics variety
* v14 migration uses `builtin.LoadTree` to get the new v4 hamt with generics to pass to `RunMigration`; the other migrations still use `LoadLegacyTree`.

So the impact of this is that when running the v14 migration we read and write the actors HAMT with the new v4 HAMT code that uses generics instead of `cbg.Deferred`.

---

## Results

`lotus-shed migrate-state 23 bafy2bzacecnpvunvyytfzmdofrdwk2jr5sf4cuzit6o6uctrzqdltqwzbwwmk` for both of them:

### Current lotus & go-state-types

migration height  4145482
old cid  bafy2bzacedpcjm7wgoq6ft3x7jdjtjdmbaqil2zjd2pa6exl2mwpv5opik4s2
new cid  bafy2bzacebpjrptlb44oaj4vyaaa5th3lvqwv5cogdfd6agzvjqr223grwhli
completed round actual (without cache), took  11.323758969s
completed round actual (with cache), took  9.755098713s

### This branch and HAMT v4

migration height  4145482
old cid  bafy2bzacedpcjm7wgoq6ft3x7jdjtjdmbaqil2zjd2pa6exl2mwpv5opik4s2
new cid  bafy2bzacebpjrptlb44oaj4vyaaa5th3lvqwv5cogdfd6agzvjqr223grwhli
completed round actual (without cache), took  10.133880646s
completed round actual (with cache), took  8.987781194s

So we're about 10.5% faster without cache and nearly 8% faster with cache.

### Flame graphs

Before:
![flame](https://github.com/user-attachments/assets/66ec0d2d-ec87-43e2-9b4b-57356548fce4)

After:
![flame2](https://github.com/user-attachments/assets/57bc6a56-885f-4c70-bb14-ec85ea0518f7)

You can see in there:

* `ForEach` on the actors tree goes from 30.9% to 28.90% (most of the action is captured under here since it nests the `Set` too
* `Set` is 11.25% vs 6.22% (this is dramatic because we get to defer serialization whereas previously we had to serialize the value into a slice and hold onto it till a flush)
* `Flush` goes up from 4.97% to 6.51% (this is where the value serialization is moved to, but it's much cheaper because we go directly to the `io.Writer` rather than a slice allocation [+GC], hold, then write)

### Profiles

(click to see full profiles)

Before:
![migrate-state cpuprofile](https://github.com/user-attachments/assets/b8c3c0bc-a253-42e9-96be-617d8e305dda)

After:
![migrate-state cpuprofile2](https://github.com/user-attachments/assets/d72a4ccb-f811-4a98-bf1d-e6841d59c51c)

These are some of the bits I was most interested in:

<img width="134" alt="Screenshot 2024-08-03 at 7 14 16 PM" src="https://github.com/user-attachments/assets/95e11655-1670-40fa-9051-6476806162d0">
vs
<img width="187" alt="Screenshot 2024-08-03 at 7 14 29 PM" src="https://github.com/user-attachments/assets/1c3b1157-e938-4750-bfcc-4ecbcab6f95f">

<br>

<img width="335" alt="Screenshot 2024-08-03 at 7 15 54 PM" src="https://github.com/user-attachments/assets/da22adb4-1f41-4733-8c81-8cfd81cb2873">
vs
<img width="292" alt="Screenshot 2024-08-03 at 7 16 08 PM" src="https://github.com/user-attachments/assets/479ae185-74cb-4454-910d-1eeae0798014">

(GC improvements not as dramatic as I hoped, but still there)

There's probably a few more tweaks that could be done to make this nicer. There's some internal mutation code in the HAMT that I think should be changed now with the generics that may speed up the path through the code for inserts and deletes.