@rvagg 2024-10-08
Bailing on this for now, as per comment @ https://github.com/whyrusleeping/cbor-gen/pull/103#issuecomment-2398852941